fix: Consolidate `is` prefixed boolean methods
